Moved AppGroup commands to AppGroups folder
